pci: Add support for PLDA's XpressRICH3 PCIe host bridge
This adds support for the XpressRich3-AXI IP block when configured
as a PCIE host bridge. Support for generating MSI interrupts is
ignored here as the platform where this driver has been developed for
does not have support for passing the interrupts from the IP into
the CPU, but uses the GICv2m functionality.

Signed-off-by: Liviu Dudau <Liviu.Dudau@arm.com>