cpufreq: arm_big_little: remove unused cpu-cluster.<n> clock name
The "cpu-cluster.<n>" used to get the cluster clock is not used by any
platform. Moreover __of_clk_get_by_name used in clk_get return error if
the "clock-names" in the DT doesn't match this string. When using DT,
it's not compulsory to specify the clock name unless there are multiple
clock input entries in the consumer.

This patch removes the unused clock string from the driver.

Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>